requirements.txt
torch
torchvision
transformers
sentence-transformers
accelerate
chromadb
pymupdf
pillow
tqdm
scipy
numpy
requests